backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
aiofiles==24.1.0
httpx==0.28.1
python-dotenv==1.0.1
passlib==1.7.4
argon2-cffi==23.1.0
